pyproject.toml
[project]
name = "iocapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apiutils-rec>=0.1.0",
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "gensim>=4.3.3",
    "loguru>=0.7.3",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
